Move OpBuilder to internal package.
