+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/internal/ReflectMethods.java
@@ -91,7 +91,6 @@ import jdk.incubator.code.extern.DialectFactory;
 import jdk.incubator.code.dialect.core.*;
 import jdk.incubator.code.dialect.java.*;
 import jdk.incubator.code.dialect.java.WildcardType.BoundKind;
-import jdk.incubator.code.extern.OpBuilder;
 
 import javax.lang.model.element.Modifier;
 import javax.tools.JavaFileObject;
